requirements.txt
python-dotenv>=1.1.0
tiktoken>=0.9.0
tqdm>=4.67.1
numpy>=2.2.6
openai>=1.84.0
jinja2>=3.1.6
langgraph>=0.4.8
zep-cloud>=2.13.1
mem0ai>=0.1.104
langmem>=0.0.27
ollama>=0.4.0
qdrant-client>=1.12.0
sentence-transformers>=3.4.0
requests>=2.25.1
psutil>=7.0.0
pandas>=2.3.0
gputil>=1.4.0